refactor(layers): clear foundation primitive edges
