+++ aragora/swarm/boss_validation.py
@@ -19,7 +19,7 @@ from pathlib import Path
 from typing import Any
 
 from aragora.swarm.mission import GateEvaluation, GateType, GateVerdict
-from aragora.utils.semantic_extraction import ExtractionProvider, extract_json_object_llm_first
+from aragora.agents.semantic_extraction import ExtractionProvider, extract_json_object_llm_first
 
 logger = logging.getLogger(__name__)
