+++ scripts/baselines/import_contracts_baseline.json
@@ -18,8 +18,6 @@
         "aragora.agents -> aragora.gauntlet",
         "aragora.agents -> aragora.nomic",
         "aragora.agents -> aragora.server",
-        "aragora.config -> aragora.observability",
-        "aragora.core_types -> aragora.debate",
         "aragora.debate -> aragora.broadcast",
         "aragora.debate -> aragora.connectors",
         "aragora.debate -> aragora.gateway",
@@ -31,7 +29,6 @@
         "aragora.debate -> aragora.server",
         "aragora.debate -> aragora.spectate",
         "aragora.debate -> aragora.workflow",
-        "aragora.errors -> aragora.server",
         "aragora.evaluation -> aragora.server",
         "aragora.events -> aragora.server",
         "aragora.evidence -> aragora.connectors",
@@ -66,11 +63,6 @@
         "aragora.swarm -> aragora.connectors",
         "aragora.swarm -> aragora.mcp",
         "aragora.swarm -> aragora.server",
-        "aragora.utils -> aragora.agents",
-        "aragora.utils -> aragora.connectors",
-        "aragora.utils -> aragora.knowledge",
-        "aragora.utils -> aragora.resilience",
-        "aragora.utils -> aragora.storage",
         "aragora.verticals -> aragora.connectors",
         "aragora.workflow -> aragora.connectors",
         "aragora.workflow -> aragora.gateway",
